requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncio==3.4.3
attrs==23.2.0
BingImageCreator==0.5.0
browser-cookie3==0.19.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
EdgeGPT @ https://github.com/jacobgelling/EdgeGPT/archive/refs/tags/0.14.1.zip
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
log2d==0.0.18
loguru==0.7.2
lz4==4.3.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==0.28.1
prompt-toolkit==3.0.43
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
python-gemini-api==2.4.11
regex==2024.5.15
requests==2.31.0
rich==13.7.1
setuptools==69.5.1
sniffio==1.3.1
socksio==1.0.0
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
wheel==0.43.0
win32-setctime==1.1.0
yarl==1.9.4
